+++ tests/testthat/test-zip.R
@@ -11,7 +11,7 @@ test_that("can compress single directory", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(tmp))
@@ -36,7 +36,7 @@ test_that("can compress single file", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(tmp))
@@ -58,7 +58,7 @@ test_that("can compress multiple files", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp1),
       zip(zipfile, basename(c(tmp1, tmp2)))
@@ -84,7 +84,7 @@ test_that("can compress multiple directories", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp1),
       zip(zipfile, basename(c(tmp1, tmp2)))
@@ -127,7 +127,7 @@ test_that("can compress files and directories", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(c(file1, tmp, file2)))
@@ -184,14 +184,14 @@ test_that("compression level is used", {
   zipfile1 <- tempfile(fileext = ".zip")
   zipfile2 <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(file),
       zip(zipfile1, basename(file), compression_level = 1)
     )
   )
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(file),
       zip(zipfile2, basename(file), compression_level = 9)
@@ -220,7 +220,7 @@ test_that("can append a directory to an archive", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(tmp))
@@ -239,7 +239,7 @@ test_that("can append a directory to an archive", {
   cat("first file2", file = file.path(tmp2, "file3"))
   cat("second file2", file = file.path(tmp2, "file4"))
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip_append(zipfile, basename(tmp2))
@@ -264,7 +264,7 @@ test_that("can append a file to an archive", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(tmp))
@@ -281,7 +281,7 @@ test_that("can append a file to an archive", {
 
   cat("first file2", file = file1 <- tempfile())
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip_append(zipfile, basename(file1))
@@ -306,7 +306,7 @@ test_that("can append files and directories to an archive", {
 
   zipfile <- tempfile(fileext = ".zip")
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(tmp))
@@ -326,7 +326,7 @@ test_that("can append files and directories to an archive", {
   cat("another", file = file.path(tmp2, "file3"))
   cat("and another", file = file.path(tmp2, "file4"))
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip_append(zipfile, basename(c(file1, tmp2)))
@@ -352,7 +352,7 @@ test_that("empty directories are archived as directories", {
   dir.create(file.path(tmp, "foo", "bar2"))
   cat("contents\n", file = file.path(tmp, "foo", "file1"))
 
-  expect_silent(
+  expect_deprecated(
     withr::with_dir(
       dirname(tmp),
       zip(zipfile, basename(tmp))